fix: nothing

+++ run_server.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 # Xiang Wang(ramwin@qq.com)
 
-rm -r _build
+if [ -d _build ]
+then
+    rm -r _build
+fi
 
-# wsl下
-# 这样都会忽略
-# sphinx-autobuild -j auto --port 18002 . _build/html/ --re-ignore _build --re-ignore .git
-# 这样git status也会刷新
-sphinx-autobuild -j auto --port 18003 . _build/html/ --re-ignore _build
+sphinx-autobuild \
+    -j auto \
+    --port 18003 \
+    . _build/html/ \
+    --re-ignore "\.mypy_cache" \
+    --re-ignore "\.git"    \
+    --re-ignore "\.*\.swp" \
+    --re-ignore "\.*\.log" \
+
